requirements.txt
instagrapi>=1.22.0
fastapi>=0.108.0
uvicorn>=0.25.0
pymongo>=4.6.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.5.0
jinja2>=3.1.2
motor>=3.3.0
loguru>=0.7.2
Pillow>=8.1.1